* Undo stack is now memory limited (to 512 megabytes). * Simplemesh now supports an indexed draw elements mode.   + which pbvj.js now uses to draw bvh nodes * Added a API for logging mesh topology changes.  The idea is to use   the new LogContext class; make an instance and set .onnew and .onkill.   Lots of API functions now accept an (always optional) LogContext instance. * Added a simple quad remesher * Added a hole filler sculpt tool * Added a quad mode for dyntopo.  It presently doesn't work well (it   generates holes). * Added a new splitEdgeSmart function for DynTopo that's a bit more   advanced then the old one (though the old one is still being used for   the edge cut tool). * Added utility isTri() isQuad() isNgon() methods to mesh.Face.
I've discovered a numerical error problem when drawing selection buffers
for meshes.  The plan is to replace mesh edit mode selection picking
with bvh picking, as it's more accurate anyway.

+++ scripts/shaders/shaders.js
@@ -832,7 +832,7 @@ void main() {
 };
 
 export let MeshIDShader = {
-  vertex : `precision mediump float;
+  vertex : `precision highp float;
   
 uniform mat4 projectionMatrix;
 uniform mat4 objectMatrix;
@@ -876,7 +876,7 @@ void main() {
   vId = id + 1.0; // + id_offset;
 }
 `,
-  fragment : `precision mediump float;
+  fragment : `precision highp float;
 
 uniform float id_offset;
 uniform float object_id;
